add a new set of tests, clean things up
